# Reduce memory used by postings offset table.