--- a/tsdb/docs/format/index.md
+++ b/tsdb/docs/format/index.md
@@ -203,9 +203,9 @@ They are used to track label index sections. They are read into memory when an i
 
 ### Postings Offset Table
 
-A postings offset table stores a sequence of postings offset entries.
+A postings offset table stores a sequence of postings offset entries, sorted by label name and value.
 Every postings offset entry holds the label name/value pair and the offset to its series list in the postings section.
-They are used to track postings sections. They are read into memory when an index file is loaded.
+They are used to track postings sections. They are partially read into memory when an index file is loaded.
 
 ```
 ┌─────────────────────┬──────────────────────┐