Rather than keeping the offset of each postings list, instead
keep the nth offset of the offset of the posting list. As postings
list offsets have always been sorted, we can then get to the closest
entry before the one we want an iterate forwards.

I haven't done much tuning on the 32 number, it was chosen to try
not to read through more than a 4k page of data.

Add a bulk interface for fetching postings. Use it to avoid having to
re-read parts of the posting offset table when querying lots of it.

For a index with what BenchmarkHeadPostingForMatchers uses RAM
for r.postings drops from 3.79MB to 80.19kB or about 48x.
Bytes allocated go down by 30%, and suprisingly CPU usage drops by
4-6% for typical queries too.

```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkHeadPostingForMatchers/n="1"-4                              35220         36959         +4.94%
BenchmarkHeadPostingForMatchers/n="1",j="foo"-4                      578798        615273        +6.30%
BenchmarkHeadPostingForMatchers/j="foo",n="1"-4                      581930        614427        +5.58%
BenchmarkHeadPostingForMatchers/n="1",j!="foo"-4                     583190        615479        +5.54%
BenchmarkHeadPostingForMatchers/i=~".*"-4                            115169232     116079317     +0.79%
BenchmarkHeadPostingForMatchers/i=~".+"-4                            155318554     146743642     -5.52%
BenchmarkHeadPostingForMatchers/i=~""-4                              60655399      49071222      -19.10%
BenchmarkHeadPostingForMatchers/i!=""-4                              49437509      39161630      -20.79%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",j="foo"-4              114361825     116101312     +1.52%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",i!="2",j="foo"-4       113697227     114773786     +0.95%
BenchmarkHeadPostingForMatchers/n="1",i!=""-4                        50452920      39238957      -22.23%
BenchmarkHeadPostingForMatchers/n="1",i!="",j="foo"-4                51290788      41008892      -20.05%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",j="foo"-4              156004011     145612748     -6.66%
BenchmarkHeadPostingForMatchers/n="1",i=~"1.+",j="foo"-4             30318091      30444402      +0.42%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!="2",j="foo"-4       155504240     148051405     -4.79%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-4     185225863     174089236     -6.01%

benchmark                                                            old allocs     new allocs     delta
BenchmarkHeadPostingForMatchers/n="1"-4                              4              5              +25.00%
BenchmarkHeadPostingForMatchers/n="1",j="foo"-4                      7              9              +28.57%
BenchmarkHeadPostingForMatchers/j="foo",n="1"-4                      7              9              +28.57%
BenchmarkHeadPostingForMatchers/n="1",j!="foo"-4                     15             16             +6.67%
BenchmarkHeadPostingForMatchers/i=~".*"-4                            100010         100012         +0.00%
BenchmarkHeadPostingForMatchers/i=~".+"-4                            200073         200047         -0.01%
BenchmarkHeadPostingForMatchers/i=~""-4                              200076         200048         -0.01%
BenchmarkHeadPostingForMatchers/i!=""-4                              200070         200041         -0.01%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",j="foo"-4              100013         100016         +0.00%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",i!="2",j="foo"-4       100017         100020         +0.00%
BenchmarkHeadPostingForMatchers/n="1",i!=""-4                        200073         200045         -0.01%
BenchmarkHeadPostingForMatchers/n="1",i!="",j="foo"-4                200075         200048         -0.01%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",j="foo"-4              200078         200053         -0.01%
BenchmarkHeadPostingForMatchers/n="1",i=~"1.+",j="foo"-4             111166         111149         -0.02%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!="2",j="foo"-4       200082         200057         -0.01%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-4     311288         311240         -0.02%

benchmark                                                            old bytes     new bytes     delta
BenchmarkHeadPostingForMatchers/n="1"-4                              264           280           +6.06%
BenchmarkHeadPostingForMatchers/n="1",j="foo"-4                      360           392           +8.89%
BenchmarkHeadPostingForMatchers/j="foo",n="1"-4                      360           392           +8.89%
BenchmarkHeadPostingForMatchers/n="1",j!="foo"-4                     520           537           +3.27%
BenchmarkHeadPostingForMatchers/i=~".*"-4                            1600547       1602997       +0.15%
BenchmarkHeadPostingForMatchers/i=~".+"-4                            24901726      17262925      -30.68%
BenchmarkHeadPostingForMatchers/i=~""-4                              24901882      17260825      -30.68%
BenchmarkHeadPostingForMatchers/i!=""-4                              24900778      17260546      -30.68%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",j="foo"-4              1600632       1601616       +0.06%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",i!="2",j="foo"-4       1600803       1600851       +0.00%
BenchmarkHeadPostingForMatchers/n="1",i!=""-4                        24900967      17259722      -30.69%
BenchmarkHeadPostingForMatchers/n="1",i!="",j="foo"-4                24900971      17259819      -30.69%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",j="foo"-4              24901861      17262861      -30.68%
BenchmarkHeadPostingForMatchers/n="1",i=~"1.+",j="foo"-4             3788509       3143975       -17.01%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!="2",j="foo"-4       24902019      17261760      -30.68%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-4     28694270      20407961      -28.88%
```

The benchmark commit is something that I hacked together to test this, and should not be merged.