@brian-brazil 2019-12-07
Okay, that'd cleaned up and in a submittable state. Comments have been addressed.