@brian-brazil 2019-12-10
The benchmarks above cover this, and I'd like to get this in as I've further significant memory improvements I'd like to do which are based on this.